# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 知涟 API
  version: 0.1.0
  description: |
    知涟是内容社区与 Feed 分发平台。本文件从 Stage 1 开始维护 API 契约，
    后续阶段会逐步补充账号、内容发布、治理审核、Feed 与交互接口。
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
paths:
  /health:
    get:
      summary: Health check
      description: 返回服务健康状态，响应必须使用统一 JSON 包装并包含 request_id。
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HealthData"
              examples:
                ok:
                  value:
                    data:
                      status: ok
                    error: null
                    request_id: req_01hzy6j6r7m9w2w5q5x8k9z0a1
  /api/users:
    post:
      summary: Register user
      description: 注册普通用户，邮箱唯一，密码以 bcrypt 哈希后保存。
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/sessions:
    post:
      summary: Login
      description: 使用邮箱和密码登录，返回 Bearer JWT。
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/sessions/current:
    delete:
      summary: Logout current session
      description: |
        注销当前会话。V1 使用 stateless JWT，服务端不维护 token blacklist，
        成功响应表示客户端应删除本地 token。
      operationId: logoutCurrentSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout acknowledged
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LogoutResult"
        "401":
          $ref: "#/components/responses/Error"
  /api/users/me:
    get:
      summary: Current user
      description: 通过 Bearer JWT 获取当前登录用户。
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update current user profile
      description: 更新当前用户公开资料。该接口是 PATCH 语义，未传字段保持原值。
      operationId: updateCurrentUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /api/users/{userId}:
    get:
      summary: Public user profile
      description: 返回用户公开资料，不包含 email、role、status、password hash 或 token state。
      operationId: getPublicUserProfile
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Public user profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PublicUser"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: API error using the unified response envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessResponse:
      type: object
      required:
        - data
        - error
        - request_id
      properties:
        data:
          nullable: true
        error:
          nullable: true
        request_id:
          type: string
          description: Request trace identifier returned in every API response.
    ErrorResponse:
      type: object
      required:
        - data
        - error
        - request_id
      properties:
        data:
          nullable: true
          example: null
        error:
          $ref: "#/components/schemas/APIError"
        request_id:
          type: string
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: title is required
    HealthData:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 6
          example: secret123
        nickname:
          type: string
          maxLength: 64
          example: Alice
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: secret123
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 64
          example: Alice
        avatar_url:
          type: string
          maxLength: 512
          example: /uploads/images/avatar.jpg
        bio:
          type: string
          maxLength: 512
          example: Go backend learner
    User:
      type: object
      required:
        - id
        - email
        - nickname
        - role
        - status
        - created_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: alice@example.com
        nickname:
          type: string
          example: Alice
        avatar_url:
          type: string
          example: ""
        bio:
          type: string
          example: ""
        role:
          type: string
          enum:
            - user
            - admin
        status:
          type: string
          enum:
            - active
            - disabled
        created_at:
          type: string
          format: date-time
    PublicUser:
      type: object
      required:
        - id
        - nickname
        - avatar_url
        - bio
        - created_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nickname:
          type: string
          example: Alice
        avatar_url:
          type: string
          example: /uploads/images/avatar.jpg
        bio:
          type: string
          example: Go backend learner
        created_at:
          type: string
          format: date-time
    LogoutResult:
      type: object
      required:
        - logged_out
      properties:
        logged_out:
          type: boolean
          example: true
    Session:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
          description: Bearer JWT token.
        user:
          $ref: "#/components/schemas/User"